Add documentation to RTC module Avoid more Wnarrowing with rtc.getSysTime Camel casing in MIRRAPins

+++ firmware/lib/ESPCamUART/ESPCamUART.cpp
@@ -50,7 +50,7 @@ SensorValue ESPCamUART::getMeasurement()
 
 void ESPCamUART::updateNextSampleTime(uint32_t sampleInterval)
 {
-    uint32_t cTime{time(nullptr)};
+    uint32_t cTime{static_cast<uint32_t>(time(nullptr))};
     uint32_t target{cTime};
     while (nextSampleTime <= cTime)
     {
